Update build item
